+++ src/neptune/new/runs_table.py
@@ -33,9 +33,9 @@ from neptune.internal.backends.api_model import (
 )
 from neptune.internal.backends.neptune_backend import NeptuneBackend
 from neptune.internal.container_type import ContainerType
-from neptune.metadata_containers.metadata_containers_table import (
+from neptune.table import (
     LeaderboardEntry,
     LeaderboardHandler,
 )
-from neptune.metadata_containers.metadata_containers_table import Table as RunsTable
-from neptune.metadata_containers.metadata_containers_table import TableEntry as RunsTableEntry
+from neptune.table import Table as RunsTable
+from neptune.table import TableEntry as RunsTableEntry
