+++ tests/unit/neptune/new/client/test_run_tables.py
@@ -28,11 +28,11 @@ from neptune.internal.backends.api_model import (
 )
 from neptune.internal.backends.neptune_backend_mock import NeptuneBackendMock
 from neptune.internal.container_type import ContainerType
-from neptune.metadata_containers.metadata_containers_table import (
+from neptune.metadata_containers.utils import DATE_FORMAT
+from neptune.table import (
     Table,
     TableEntry,
 )
-from neptune.metadata_containers.utils import DATE_FORMAT
 from tests.unit.neptune.new.client.abstract_tables_test import AbstractTablesTestMixin
